+++ upload/app/group/App/Class/Controller/Grouptopicadmin_/DeletetopicController_.php
@@ -18,24 +18,38 @@ class DeletetopicController extends Controller{
 
 		if(is_array($arrGrouptopics)){
 			foreach($arrGrouptopics as $nGrouptopic){
-				$oGrouptopicMeta=GrouptopicModel::M();
-				$oGrouptopicMeta->deleteWhere(array('grouptopic_id'=>$nGrouptopic));
-				
-				if($oGrouptopicMeta->isError()){
-					$this->E($oGrouptopicMeta->getErrorMessage());
-				}
+				// 帖子回收站功能开启
+				if($GLOBALS['_cache_']['group_option']['group_deletetopic_recyclebin']==1){
+					$oGrouptopic=GrouptopicModel::F('grouptopic_id=?',$nGrouptopic)->getOne();
 
-				// 删除主题关联的回帖
-				$oGrouptopiccommentMeta=GrouptopiccommentModel::M();
-				$oGrouptopiccommentMeta->deleteWhere(array('grouptopic_id'=>$nGrouptopic));
-				
-				if($oGrouptopiccommentMeta->isError()){
-					$this->E($oGrouptopiccommentMeta->getErrorMessage());
+					if(!empty($oGrouptopic['grouptopic_id'])){
+						$oGrouptopic->grouptopic_status='0';
+						$oGrouptopic->save(0,'update');
+
+						if($oGrouptopic->isError()){
+							$this->E($oGrouptopic->getErrorMessage());
+						}
+					}
+				}else{
+					$oGrouptopicMeta=GrouptopicModel::M();
+					$oGrouptopicMeta->deleteWhere(array('grouptopic_id'=>$nGrouptopic));
+					
+					if($oGrouptopicMeta->isError()){
+						$this->E($oGrouptopicMeta->getErrorMessage());
+					}
+
+					// 删除主题关联的回帖
+					$oGrouptopiccommentMeta=GrouptopiccommentModel::M();
+					$oGrouptopiccommentMeta->deleteWhere(array('grouptopic_id'=>$nGrouptopic));
+					
+					if($oGrouptopiccommentMeta->isError()){
+						$this->E($oGrouptopiccommentMeta->getErrorMessage());
+					}
 				}
 			}
 		}
 
-		$sGroupurl=Group_Extend::getGroupurl($oGroup);
+		$sGroupurl=Group_Extend::getGroupurl($nGroupid);
 
 		$this->A(array('group_id'=>$nGroupid,'group_url'=>$sGroupurl),Dyhb::L('删除主题成功','Controller/Grouptopicadmin'));
 	}
